+++ thewalrus/tests/test_decompositions.py
@@ -323,6 +323,14 @@ def test_takagi_error():
     with pytest.raises(ValueError, match="The input matrix is not square"):
         takagi(A)
 
+def test_takagi_sepcific_matrix():
+    """Test the takagi decomposition works well for a specific matrix that was not deecomposed accuratelyin a previous version.
+    See more info in PR #393 (https://github.com/XanaduAI/thewalrus/pull/393)"""
+    A = np.load('test_matrix_for_takagi.npy')
+    d, U = takagi(A)
+    assert np.allclose(A, U @ np.diag(d) @ U.T)
+    assert np.allclose(U @ np.conjugate(U).T, np.eye(len(U)))
+    assert np.all(d >= 0)
 
 def test_real_degenerate():
     """Verify that the Takagi decomposition returns a matrix that is unitary and results in a